grid file rename, dependencies updated
